Add more content to pyldt.py.
Add copy_raw() and biascombine() methods to _Images class, which will
inherit to the LMI and DeVeny classes.

Added an import command to __init__.py to make the package work in
an expected manner (i.e. when importing the package, all methods and
classes from the various files are available for call.