Have basic LMI calibration pipeline working
I have moved over the remaining pieces from use_ccdproc.py to pyldt.py.
The user-facing methods in the LMI class do the basic calibration steps, with a
wrapper method process_all() to run the user-facing methods in order on the
directory used to initialize the LMI class.  When initializing the class,
the user may specify the bin_factor, and the class will then restrict its
attention to files with that binning factor only.  Creating a new instance of
the LMI class with a different bin_factor will allow the user to access those
images with the new binning factor.

Next steps: bring the DeVeny class up to approximately the same level of
functionality.  Also, document, document, document.