Add HISTORY to FITS headers, move flat combine to LMI
This should have been two separate commits, but I forgot.  The first bit
is the addition of various HISTORY fields to the FITS headers to indicate the
pedigree of the current file.

The second was moving the _flatcombine() method from _Images to LMI and
replacing the extant flat_combine() method.  This is because the filter-
searching algorithm doesn't really work for DeVeny the way it does for LMI.