In imcombine(), unpack single-item tuple *infiles, if specified
Using the * operator for infiles allows for the function to be called without a
list, but rather with the inlist= keyword selected.  This operator, however,
returns a tuple that must be unpacked in order to be used.  This change checks
for the proper input and unpacks the tuple as required.