First stab at twodspec.longslit.response functionality
This is a reasonable first stab at a GUI/interactive functionality for the
response function used to flatten spectral flat field images.  The user can
select a fit function (only 1 implemented currently) and try different orders
of that function to see the effect on the resulting spectrum.

Appliction of the fit to division of the original image is not yet implemented.

There is lots of cruft in here at the moment related to debugging, but that will
be sorted out.

	modified:   __init__.py
	modified:   twodspec.py
	modified:   utils.py