Polish cubic spline fitting
In utils.py, overhaul the cubic spline fitting to include the "not-a-knot"
boundary condition (essentially removing the first and last knots to allow for
a smoother end segment), and finesse the fitting procedure (including checing
for errors).  The routine spline3() now returns the fitted spline, knot
location, and any error message to be displayed.

	modified:   twodspec.py
	modified:   utils.py